Move enricher plugins into separate files
