+++ addons/input_module/localfiles/enricher.py
@@ -1,14 +1,11 @@
 import logging
 import time
 import threading
-import os
-import abc
-import requests
-from typing import Dict, List, Optional, Any, Set
-from PIL import Image
-import io
-import musicbrainzngs
 import queue
+
+from .enricher_plugin import EnricherPlugin
+from .musicbrainz_plugin import MusicBrainzPlugin
+from .wikidata_plugin import WikidataPlugin
 
 logger = logging.getLogger(__name__.split(".")[-1])
 
@@ -20,412 +17,6 @@ musicbrainz_logger.setLevel(logging.WARNING)
 _enricher_thread = None
 _enricher_instance = None
 _enricher_queue = queue.Queue()
-
-
-class EnricherPlugin(abc.ABC):
-    """Base class for enricher plugins"""
-
-    @abc.abstractmethod
-    def can_enrich_artist(self) -> bool:
-        """Whether this plugin can enrich artist metadata"""
-        pass
-
-    @abc.abstractmethod
-    def can_enrich_album(self) -> bool:
-        """Whether this plugin can enrich album metadata"""
-        pass
-
-    @abc.abstractmethod
-    def can_enrich_track(self) -> bool:
-        """Whether this plugin can enrich track metadata"""
-        pass
-
-    @abc.abstractmethod
-    def enrich_artist(self, artist: Dict) -> Optional[Dict]:
-        """Enrich artist metadata"""
-        pass
-
-    @abc.abstractmethod
-    def enrich_album(self, album: Dict) -> Optional[Dict]:
-        """Enrich album metadata"""
-        pass
-
-    @abc.abstractmethod
-    def enrich_track(self, track: Dict) -> Optional[Dict]:
-        """Enrich track metadata"""
-        pass
-
-
-class MusicBrainzPlugin(EnricherPlugin):
-    """MusicBrainz metadata enrichment plugin"""
-
-    def __init__(self, config, db_manager):
-        self.config = config
-        self.db_manager = db_manager
-        self.threshold = config["enricher.plugins.musicbrainz.match_threshold"]
-        self.user_agent = config["enricher.plugins.musicbrainz.user_agent"]
-
-        # Set up MusicBrainz API
-        musicbrainzngs.set_useragent(
-            self.user_agent.split("/")[0],
-            self.user_agent.split("/")[1].split(" ")[0],
-            self.user_agent.split(" ", 1)[1].strip("()"),
-        )
-
-        # Rate limiting
-        self.last_request_time = 0
-        self.request_interval = 1 / 3  # 3 requests per second
-
-    def _wait_for_rate_limit(self):
-        """Wait to respect rate limits"""
-        now = time.time()
-        elapsed = now - self.last_request_time
-
-        if elapsed < self.request_interval:
-            sleep_time = self.request_interval - elapsed
-            time.sleep(sleep_time)
-
-        self.last_request_time = time.time()
-
-    def can_enrich_artist(self) -> bool:
-        return True
-
-    def can_enrich_album(self) -> bool:
-        return True
-
-    def can_enrich_track(self) -> bool:
-        return True
-
-    def enrich_artist(self, artist: Dict) -> Optional[Dict]:
-        """Enrich artist metadata with MusicBrainz data"""
-        try:
-            if artist["id"] == "unknown_artist":
-                return None
-
-            logger.debug(f"Enriching artist: {artist['name']}")
-
-            # Search for artist in MusicBrainz
-            self._wait_for_rate_limit()
-            result = musicbrainzngs.search_artists(artist["name"])
-
-            if not result["artist-list"]:
-                logger.debug(f"No MusicBrainz match found for artist: {artist['name']}")
-                return None
-
-            # Find best match
-            best_match = None
-            best_score = 0
-
-            for mb_artist in result["artist-list"]:
-                score = int(mb_artist.get("ext:score", 0))
-                if score > best_score:
-                    best_score = score
-                    best_match = mb_artist
-
-            if best_score < self.threshold:
-                logger.debug(
-                    f"No good MusicBrainz match for artist: {artist['name']} (best score: {best_score})"
-                )
-                return None
-
-            # Get more details from artist
-            artist_mbid = best_match["id"]
-            self._wait_for_rate_limit()
-            mb_artist_details = musicbrainzngs.get_artist_by_id(
-                artist_mbid, includes=["url-rels"]
-            )
-            mb_artist_data = mb_artist_details["artist"]
-
-            # Update artist data
-            updates = {"mbid": artist_mbid, "enriched": 1}
-
-            # Return data for further enrichment if needed
-            return {"updates": updates, "mbid": artist_mbid}
-
-        except Exception as e:
-            logger.error(f"Error enriching artist {artist['name']}: {str(e)}")
-            return None
-
-    def enrich_album(self, album: Dict) -> Optional[Dict]:
-        """Enrich album metadata with MusicBrainz data"""
-        try:
-            if album["id"] == "unknown_album":
-                return None
-
-            logger.debug(f"Enriching album: {album['title']} by {album['artist_name']}")
-
-            # Search for album in MusicBrainz
-            self._wait_for_rate_limit()
-            result = musicbrainzngs.search_releases(
-                album["title"], artistname=album["artist_name"]
-            )
-
-            if not result["release-list"]:
-                logger.debug(
-                    f"No MusicBrainz match found for album: {album['title']} by {album['artist_name']}"
-                )
-                return None
-
-            # Find best match
-            best_match = None
-            best_score = 0
-
-            for mb_release in result["release-list"]:
-                score = int(mb_release.get("ext:score", 0))
-                if score > best_score:
-                    best_score = score
-                    best_match = mb_release
-
-            if best_score < self.threshold:
-                logger.debug(
-                    f"No good MusicBrainz match for album: {album['title']} (best score: {best_score})"
-                )
-                return None
-
-            # Get more details about the release
-            release_mbid = best_match["id"]
-            self._wait_for_rate_limit()
-            mb_release_details = musicbrainzngs.get_release_by_id(
-                release_mbid, includes=["recordings", "artist-credits", "genres"]
-            )
-            mb_release_data = mb_release_details["release"]
-
-            # Update album data
-            updates = {"mbid": release_mbid, "enriched": 1}
-
-            # Add genre if available
-            if "genre-list" in mb_release_data and mb_release_data["genre-list"]:
-                updates["genre"] = mb_release_data["genre-list"][0]["name"]
-
-            # Add year if available
-            if "date" in mb_release_data:
-                try:
-                    updates["year"] = int(mb_release_data["date"].split("-")[0])
-                except (ValueError, IndexError):
-                    pass
-
-            # Return data for further enrichment if needed
-            return {"updates": updates, "mbid": release_mbid}
-
-        except Exception as e:
-            logger.error(f"Error enriching album {album['title']}: {str(e)}")
-            return None
-
-    def enrich_track(self, track: Dict) -> Optional[Dict]:
-        """Enrich track metadata with MusicBrainz data"""
-        try:
-            logger.debug(
-                f"Enriching track: {track['title']} from {track['album_title']}"
-            )
-
-            # Search for recording in MusicBrainz
-            self._wait_for_rate_limit()
-            result = musicbrainzngs.search_recordings(
-                track["title"],
-                artistname=track["artist_name"],
-                release=track["album_title"],
-            )
-
-            if not result["recording-list"]:
-                logger.debug(f"No MusicBrainz match found for track: {track['title']}")
-                return None
-
-            # Find best match
-            best_match = None
-            best_score = 0
-
-            for mb_recording in result["recording-list"]:
-                score = int(mb_recording.get("ext:score", 0))
-                if score > best_score:
-                    best_score = score
-                    best_match = mb_recording
-
-            if best_score < self.threshold:
-                logger.debug(
-                    f"No good MusicBrainz match for track: {track['title']} (best score: {best_score})"
-                )
-                return None
-
-            # Update track data
-            recording_mbid = best_match["id"]
-            updates = {"mbid": recording_mbid, "enriched": 1}
-
-            # Return data for further enrichment if needed
-            return {"updates": updates, "mbid": recording_mbid}
-
-        except Exception as e:
-            logger.error(f"Error enriching track {track['title']}: {str(e)}")
-            return None
-
-
-class WikidataPlugin(EnricherPlugin):
-    """Wikidata enrichment plugin for artist images"""
-
-    def __init__(self, config, db_manager):
-        self.config = config
-        self.db_manager = db_manager
-        self.artwork_path = config["artwork_path"]
-        self.session = requests.Session()
-
-        # Rate limiting
-        self.last_request_time = 0
-        self.request_interval = 1  # 1 request per second
-
-    def _wait_for_rate_limit(self):
-        """Wait to respect rate limits"""
-        now = time.time()
-        elapsed = now - self.last_request_time
-
-        if elapsed < self.request_interval:
-            sleep_time = self.request_interval - elapsed
-            time.sleep(sleep_time)
-
-        self.last_request_time = time.time()
-
-    def can_enrich_artist(self) -> bool:
-        return True
-
-    def can_enrich_album(self) -> bool:
-        return False
-
-    def can_enrich_track(self) -> bool:
-        return False
-
-    def enrich_artist(self, artist: Dict) -> Optional[Dict]:
-        """Enrich artist with image from Wikidata via MusicBrainz"""
-        try:
-            if not artist.get("mbid"):
-                logger.debug(f"No MusicBrainz ID for artist: {artist['name']}")
-                return None
-
-            # Get MusicBrainz artist with relations
-            artist_mbid = artist["mbid"]
-
-            self._wait_for_rate_limit()
-            mb_result = musicbrainzngs.get_artist_by_id(
-                artist_mbid, includes=["url-rels"]
-            )
-            mb_artist = mb_result["artist"]
-
-            # Look for Wikidata relation
-            wikidata_url = None
-            for relation in mb_artist.get("url-relation-list", []):
-                if relation.get("type") == "wikidata":
-                    wikidata_url = relation["target"]
-                    break
-
-            if not wikidata_url:
-                logger.debug(f"No Wikidata link for artist: {artist['name']}")
-                return None
-
-            # Extract Wikidata ID (Q number)
-            wikidata_id = wikidata_url.split("/")[-1]
-
-            # Query Wikidata API for P18 (image) property
-            self._wait_for_rate_limit()
-            wikidata_api_url = "https://www.wikidata.org/w/api.php"
-            params = {
-                "action": "wbgetclaims",
-                "entity": wikidata_id,
-                "property": "P18",  # P18 is the image property
-                "format": "json",
-            }
-
-            response = self.session.get(wikidata_api_url, params=params)
-            data = response.json()
-
-            # Extract image filename from response
-            image_claims = data.get("claims", {}).get("P18", [])
-            if not image_claims:
-                logger.debug(f"No image found in Wikidata for artist: {artist['name']}")
-                return None
-
-            image_filename = image_claims[0]["mainsnak"]["datavalue"]["value"]
-
-            # Format the image URL (Wikimedia Commons)
-            # MD5 hash the filename for the URL path
-            import hashlib
-
-            filename_md5 = hashlib.md5(
-                image_filename.replace(" ", "_").encode("utf-8")
-            ).hexdigest()
-
-            image_url = f"https://upload.wikimedia.org/wikipedia/commons/{filename_md5[0]}/{filename_md5[0:2]}/{image_filename.replace(' ', '_')}"
-
-            # Download the image
-            self._wait_for_rate_limit()
-            image_response = self.session.get(image_url)
-            if image_response.status_code != 200:
-                logger.error(
-                    f"Failed to download image for artist {artist['name']}: {image_response.status_code}"
-                )
-                return None
-
-            # Save the image in different sizes
-            image_data = image_response.content
-            self._save_images(image_data, artist["id"], "artist")
-
-            # Update artist data
-            updates = {
-                "image_url": artist["id"],
-            }
-
-            return {"updates": updates}
-
-        except Exception as e:
-            logger.error(
-                f"Error enriching artist {artist['name']} with Wikidata image: {str(e)}"
-            )
-            return None
-
-    def enrich_album(self, album: Dict) -> Optional[Dict]:
-        """Placeholder for album enrichment - not implemented"""
-        return None
-
-    def enrich_track(self, track: Dict) -> Optional[Dict]:
-        """Placeholder for track enrichment - not implemented"""
-        return None
-
-    def _save_images(self, image_data: bytes, entity_id: str, entity_type: str):
-        """Save artwork images in different sizes"""
-        try:
-            img = Image.open(io.BytesIO(image_data))
-
-            # Create the directory if it doesn't exist
-            dir_path = os.path.join(self.artwork_path, entity_type)
-            os.makedirs(dir_path, exist_ok=True)
-
-            # Convert to RGB if needed (for PNG, etc.)
-            if img.mode != "RGB":
-                img = img.convert("RGB")
-
-            # Save thumbnail (50x50)
-            thumbnail = img.copy()
-            thumbnail.thumbnail((50, 50), Image.LANCZOS)
-            thumbnail.save(
-                os.path.join(dir_path, f"{entity_id}_thumbnail.jpg"), "JPEG", quality=90
-            )
-
-            # Save small (230x230)
-            small = img.copy()
-            small.thumbnail((230, 230), Image.LANCZOS)
-            small.save(
-                os.path.join(dir_path, f"{entity_id}_small.jpg"), "JPEG", quality=90
-            )
-
-            # Save large (600x600 or original if smaller)
-            large = img.copy()
-            large.thumbnail((600, 600), Image.LANCZOS)
-            large.save(
-                os.path.join(dir_path, f"{entity_id}_large.jpg"), "JPEG", quality=90
-            )
-
-            return True
-        except Exception as e:
-            logger.error(
-                f"Error saving artwork for {entity_type} {entity_id}: {str(e)}"
-            )
-            return False
 
 
 class MetadataEnricher:
